requirements.txt
customtkinter>=5.2.0
opencv-python>=4.10.0.84
Pillow>=11.0.0
pyautogui>=0.9.54
numpy>=2.1.0
keyboard>=0.13.5
pynput>=1.7.6
